agent/requirements.txt
requests>=2.31
cloudinary>=1.36
mss>=9.0
pynput>=1.7
psutil>=5.9
Pillow>=10.0
python-dotenv>=1.0
